backend/requirements/base.txt
django
django-environ
djangorestframework
django-cors-headers
django-filter
django-autoslug
django-countries
django-phonenumber-field
phonenumbers
drf-yasg
Pillow
argon2-cffi
pytz
django-taggit
celery
redis
flower
django-celery-email
django-allauth
djangorestframework-simplejwt
dj-rest-auth
django-filter
django-elasticsearch-dsl==7.3.0
django-elasticsearch-dsl-drf==0.22.5
django-storages
boto3
requests